llmlocal/requirements.txt
openai
foundry-local-sdk
